+++ src/index.ts
@@ -1,4 +1,4 @@
 export { version } from "../package.json";
-export * from "./args";
+export * from "./args.js";
 export * from "./build";
 export * from "./utils";