+++ src/utils.ts
@@ -61,9 +61,9 @@ export interface ExternalPluginOptions {
 }
 
 export function external(options: ExternalPluginOptions = {}): Plugin {
-  const CommonExts =
-    // cspell:disable-next-line
-    /\.(css|less|sass|scss|styl|stylus|pcss|postcss|json|png|jpe?g|gif|svg|ico|webp|avif|mp4|webm|ogg|mp3|wav|flac|aac|woff2?|eot|ttf|otf|wasm)$/;
+  // cspell:disable-next-line
+  // prettier-ignore
+  const CommonExts = /\.(css|less|sass|scss|styl|stylus|pcss|postcss|sss|json|apng|png|jpe?g|jfif|pjpeg|pjp|gif|svg|ico|webp|avif|mp4|webm|ogg|mp3|wav|flac|aac|opus|mov|m4a|vtt|woff2?|eot|ttf|otf|webmanifest|pdf|txt|wasm)(\?.*)?$/;
 
   const filter = options.filter ?? /^[\w@][^:]/;
   const callback = options.onResolve ?? noop;