+++ test/include.ts
@@ -1,4 +1,3 @@
-// node dist/bin.mjs --include:has --include:function-bind test/include.ts
-// @ts-ignore - it is in the node_modules
-import has from "has";
-console.log(has);
+// node dist/bin.js --include:preact test/include.ts
+import { h } from "preact";
+console.log(h);