+++ test/plugin.ts
@@ -1,4 +1,4 @@
-// node dist/bin.mjs -p:./test/plugin.ts test/index.ts
+// node dist/bin.js -p:./test/plugin.ts test/index.ts
 import { Plugin } from "esbuild";
 import { readFile } from "fs/promises";