+++ test/throw.ts
@@ -1,1 +1,2 @@
+// The first line's error offset is incorrect because we inject shims.
 throw new Error("hello world!" as string);