+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - **feat**: Shim `import.meta.{dirname,filename}`.
+- **refactor**: Use `"type": "module"` and change main file extension to `.js`.
 
 ## 0.10.5